pyproject.toml
[project]
name = "aitranscript"
version = "0.1.0"
description = "Voice transcription tool with AI-powered cleanup capabilities"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AiTranscript Team" }
]
keywords = ["transcription", "whisper", "ai", "streamlit", "voice-to-text"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web Framework
    "streamlit>=1.29.0",
    
    # Transcription & Audio Processing
    "openai-whisper>=20231117",
    "ffmpeg-python>=0.2.0",
    
    # YouTube Integration
    "youtube-transcript-api>=0.6.1",
    
    # AI Services
    "openai>=1.6.1",
    
    # Streamlit Extensions
    "audio-recorder-streamlit>=0.0.5",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "validators>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/aitranscript"
Repository = "https://github.com/yourusername/aitranscript"
Issues = "https://github.com/yourusername/aitranscript/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true